renamed library into rpg_single_board_io
