+++ launch/gpio_interrupt_example.launch
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <launch>
 
-	<node pkg="rpg_odroid_io" type="gpio_interrupt_example" name="gpio_interrupt_example" output="screen">
+	<node pkg="rpg_single_board_io" type="gpio_interrupt_example" name="gpio_interrupt_example" output="screen">
 
 	</node>
