+++ launch/voltage_reader.launch
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <launch>
 
-	<node pkg="rpg_odroid_io" type="voltage_reader" name="voltage_reader" output="screen">
+	<node pkg="rpg_single_board_io" type="voltage_reader" name="voltage_reader" output="screen">
 		<param name="board_name" value="odroid" />
 		<param name="adc_id" value="0" />
 		<param name="read_voltage_frequency" value="10" />
